requirements.txt
flask
openai-whisper
torch
transformers
ffmpeg-python